# backend/src/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Disaster Response Coordination Platform API
  description: >
    Production-grade RESTful API and Real-Time Gateway for disaster crisis management,
    geospatial resource allocation (PostGIS ST_DWithin), resilient external social reports
    stream aggregation, and Redis Cache-Aside acceleration.
  version: 1.0.0
  contact:
    name: Disaster Relief Engineering Team
    email: dev@relief.io
servers:
  - url: http://localhost:3000
    description: Local Docker Development Server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter your JWT Bearer token obtained from /api/auth/login
  schemas:
...
    DisasterInput:
      type: object
      required:
        - title
        - description
      properties:
        title:
          type: string
          example: Massive Flash Flood in Downtown Miami
        description:
          type: string
          example: Category 4 storm surge causing rising waters near Miami, FL. Roads impassable.
        location_name:
          type: string
          example: Miami, FL
        latitude:
          type: number
          example: 25.7617
        longitude:
          type: number
          example: -80.1918
        tags:
          type: array
          items:
            type: string
          example: [flood, hurricane, evacuation]
        status:
          type: string
          enum: [active, resolved, monitoring]
          example: active
...
paths:
...
  /disasters:
    get:
      tags:
        - Disasters
      summary: List all disaster incidents
      description: Retrieve paginated disasters with optional filtering by status and tag.
      parameters:
        - in: query
          name: tag
          schema:
            type: string
          description: Filter by specific tag (e.g. flood, wildfire)
        - in: query
          name: status
          schema:
            type: string
            enum: [active, resolved, monitoring]
          description: Filter by status
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        200:
          description: List of disasters

    post:
      tags:
        - Disasters
      summary: Create new disaster incident
      description: >
        Creates a new disaster incident. If coordinates are omitted, the system performs
        automatic NLP location extraction from the description text and coordinates resolution.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DisasterInput'
      responses:
        201:
          description: Disaster incident created successfully
        400:
          description: Validation error
        401:
          description: Unauthorized
        403:
          description: Forbidden (insufficient role permissions)
...
  /jobs/{jobId}:
    get:
      tags:
        - Jobs
      summary: Get asynchronous job execution status
      description: Query the real-time execution state, processing telemetry, and results of an enqueued background job.
      parameters:
        - in: path
          name: jobId
          required: true
          schema:
            type: string
            format: uuid
          description: Job unique identifier
      responses:
        200:
          description: Job status and telemetry retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  type:
                    type: string
                    example: sync_external_reports
                  disasterId:
                    type: string
                    format: uuid
                  status:
                    type: string
                    enum: [queued, processing, completed, failed]
                  progress:
                    type: integer
                    example: 100
                  result:
                    type: object
                  createdAt:
                    type: string
                  updatedAt:
                    type: string
        404:
          description: Job not found